Movement in not a straight line fixed
